// src/types/Behavior.ts
export interface BehaviorParams {
  geoPoints?: [number, number][];
  laneWidthM?: number;
  defendRadiusM?: number;
  separationDistanceM?: number;
  lineYawDeg?: number;
  robotYawDeg?: number;
  desiredFinalYawDeg?: number;
  rallyRadiusM?: number;
  outerRadius?: number;
  innerRadius?: number;
  surroundRadiusM?: number;
}